requirements.txt
asgiref==3.7.2
# boto3==1.28.15
django-cors-headers==4.2.0
django-environ==0.10.0
# django-storages==1.13.2
Django==4.2.3
djangorestframework-simplejwt==5.2.2
djangorestframework==3.14.0
drf-yasg==1.21.7
faster-whisper==0.7.1
langchain==0.0.249
openai==0.27.8
# pydub==0.25.1
# setuptools==68.0.0
shortuuid==1.0.11
# speechrecognition==3.10.0
sqlparse==0.4.4
mysqlclient==2.1.1
django-environ==0.10.0
faker==19.3.1